Refactor `lower` and reorganize tests I split the `lower` modules into two distinct ones: * a sub module `ir/lowering` which deals with translating   `chalk-parse` AST to chalk constructs * a new module `rules` which deals with outputting program   clauses and doing various things around it like default   impls and WF requirements Also I split the former `lower/test.rs` into multiple files and moved some useful macros and functions in `test_util.rs`.
